Create link_for_website
In this file i add the link for the website i've created last year for a project for a lesson of mine showing a green idea about a digital product. My idea was the creation of an AI tool that will help handicap people to have easy access to the web world.